% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eln4rmd.R
\name{elnjp_md}
\alias{elnjp_md}
\title{make new Japanese e-labnotebook in Markdown}
\usage{
elnjp_md(add_name = FALSE, replace_date = FALSE, rc = FALSE)
}
\arguments{
\item{add_name}{If you want to add the name after the date as the file name,
add this argument.}

\item{replace_date}{If you want to use the specified date as the file name
instead of today's date, add this argument.}

\item{rc}{If you are using Research Compendium of senshuRmd,
you can create a e-labnotebook file in the "labnote" directory from the current directory.
In that case, please set rc to TURE.}
}
\description{
make new Japanese e-labnotebook in Markdown
}
