% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eln4rmd.R
\name{render_elnjp_pdf}
\alias{render_elnjp_pdf}
\title{render function for Japanese e-labnotebook in PDF}
\usage{
render_elnjp_pdf(Rmd_file)
}
\arguments{
\item{Rmd_file}{file name of R Markdown file}
}
\description{
render function for Japanese e-labnotebook in PDF
}
