% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis.R
\name{senshu_thesis}
\alias{senshu_thesis}
\title{R Markdown output formats Template for thesis at Department of Psychology, Senshu University
Format for thesis at Department of Psychology, Senshu University}
\usage{
senshu_thesis()
}
\description{
R Markdown output formats Template for thesis at Department of Psychology, Senshu University
Format for thesis at Department of Psychology, Senshu University
}
