% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis.R
\name{set_analysis}
\alias{set_analysis}
\title{R Markdown output formats template for analysis at Department of Psychology, Senshu University}
\usage{
set_analysis(Rmd_file, rc = TRUE)
}
\arguments{
\item{Rmd_file}{file name of R Markdown file}

\item{rc}{If you are using Research Compendium of senshuRmd,
you can create a Rmd file in the "analysis" directory from the current directory.}
}
\description{
R Markdown output formats template for analysis at Department of Psychology, Senshu University
}
