% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbat.R
\name{set_cbat}
\alias{set_cbat}
\title{Set template files for CBAT}
\usage{
set_cbat(name_repository = "name_repository", jsPsych_version = "6.3.1")
}
\arguments{
\item{name_repository}{Set name of GitHub repository}

\item{jsPsych_version}{If you set a specific version number of jsPsych,
set_jsPsych prepare a file with that version of jsPsych.}
}
\description{
Set template files for CBAT
}
\examples{
# set_cbat("stroop")
}
