% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis.R
\name{set_rc}
\alias{set_rc}
\title{Set Research Compendium of Department of Psychology, Senshu University}
\usage{
set_rc()
}
\description{
Set Research Compendium of Department of Psychology, Senshu University
}
\examples{
# set_rc()
}
