% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eln4rmd.R
\name{up_elnjp_osf}
\alias{up_elnjp_osf}
\title{upload Japanese e-labnotebook to OSF}
\usage{
up_elnjp_osf(
  add_name = FALSE,
  replace_date = FALSE,
  eln_osf = FALSE,
  rc_osf = FALSE
)
}
\arguments{
\item{add_name}{If you want to add the name after the date as the file name,
add this argument.}

\item{replace_date}{If you want to use the specified date as the file name
instead of today's date, add this argument.}

\item{eln_osf}{URL of pdf directory in OSF}

\item{rc_osf}{If you are using Research Compendium of senshuRmd,
you can create a e-labnotebook file in the "labnote" directory from the current directory.
In that case, please set rc to TURE.}
}
\description{
upload Japanese e-labnotebook to OSF
}
