% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis.R
\name{senshu_thesis}
\alias{senshu_thesis}
\title{R Markdown output formats Template for thesis at Department of Psychology, Senshu University}
\usage{
senshu_thesis(Rmd_file, Bib_file)
}
\arguments{
\item{Rmd_file}{file name of R Markdown file}

\item{Bib_file}{file name of Bib file}
}
\description{
R Markdown output formats Template for thesis at Department of Psychology, Senshu University
}
