% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbat.R
\name{set_cbat}
\alias{set_cbat}
\title{Set template files for CBAT}
\usage{
set_cbat(task_name = "task_name", jsPsych_version = "7.1.2", use_rc = TRUE)
}
\arguments{
\item{task_name}{name of task}

\item{jsPsych_version}{If you set a specific version number of jsPsych,
set_jsPsych prepare a file with that version of jsPsych.}

\item{use_rc}{If you don"t use the RC, set FALSE.}
}
\description{
Set template files for CBAT
}
\examples{
# set_cbat("stroop")
}
