% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis.R
\name{set_rc}
\alias{set_rc}
\title{Set Research Compendium of senshu}
\usage{
set_rc(file_name = "paper")
}
\arguments{
\item{file_name}{file name and directory name of RMarkdown}
}
\value{
Make directories for Research Compendium of CCP Lab and
R Markdown file for thesis at Department of Psychology, Senshu University
}
\description{
Set Research Compendium of senshu
}
\examples{
# set_rc()
}
